pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-email-triage"
version = "1.0.0"
description = "A real-world environment for an AI agent to learn how to manage an inbox and schedule calendar events."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.20.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "gradio>=4.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
]

[tool.setuptools]
packages = ["."]
